pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "Cython", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "zoom_particle_load_creator"
description = "Make a particle load for Virgo Consortium-style zoom-in simulations."
readme = "README.md"
requires-python = ">=3.7"
license = {text = "BSD 3-Clause License"}
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: BSD 3-Clause License",
]
dependencies = [
    "numpy<2.0",
    "h5py",
    "pyaml",
    "mpi4py",
    "astropy",
    "matplotlib",
    "scipy"
]
version = "0.0.2"
authors = [
  { name="Stuart McAlpine", email="stuart.mcalpine@fysik.su.se" },
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
zoom-particle-load-creator = "particle_load.scripts.generate_particle_load:main"

[tool.setuptools.package-data]
"particle_load.glass_files" = ["ascii_glass*"]
"particle_load.template_files" = ["*", "*/**"]
